contributing/samples/state_memory_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
kafka-python==2.0.2
pydantic==2.4.2